requirements.txt
# Core dependencies
boto3>=1.34.0
qdrant-client>=1.7.0
tqdm>=4.65.0
numpy>=1.24.0

# AWS CLI
awscli

# Embeddings and ML
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.7.0
accelerate>=0.26.0

# Fix for timm compatibility issue
timm>=1.0.3

# LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain_huggingface>=0.0.1

# Optional utilities
pandas
hf_transfer
pyyaml
